added weight loading and freezing

+++ datasets/base/common/memory_mapped/constructor.py
@@ -25,9 +25,9 @@ def memory_mapped_constructor_common_preliminary_works(base_dataset: dict, base_
 
         bounding_box_data_type = base_dataset['context']['bounding_box_data_type']
         if bounding_box_data_type == 'int':
-            bounding_box_data_type = np.int
+            bounding_box_data_type = int
         elif bounding_box_data_type == 'float':
-            bounding_box_data_type = np.float
+            bounding_box_data_type = float
         else:
             raise RuntimeError(f"Unknown value {bounding_box_data_type}")
     else:
